+++ src/main/java/com/matlb/service/UserService.java
@@ -22,13 +22,13 @@ public interface UserService {
     public User saveUser(User user);
     public Subscriber saveSubscriber(Subscriber subscriber);
     public User findUserByEmail(String email);
-    public User findByMobileNumber(Integer phoneNumber);
+    public User findByMobileNumber(Long phoneNumber);
     public User findUserByEmailIdAndAuthToken(String emailId, String token);
     public Subscriber findSubscriberByEmail(String email);
     public void sendMail(String emailId);
     public User authenticateSignInUser(String email , String userToken);
     public String createUserAuthToken(String token , String email);
-    public UserResponse updateMobileNumber(User user , Integer mobileNumber);
+    public UserResponse updateMobileNumber(User user , Long mobileNumber);
     public UserResponse updateGCMToken(User user);
     public FriendsPresentResponse verifyMobileNumbers(FriendsPresentRequest friendsPresentRequest);
